test for the MPPT hexMessage parser. reorganization of functions
