+++ MTU_STM32/Core/Src/MTU/MPPT_parsing.c
@@ -6,102 +6,7 @@
  */
 
 
-#include "MTU/sensorParser.h"
-#include "stdio.h"
-#include "stdlib.h"
-#include "ctype.h"
-
-bool nmeaChecksumCompare(uint8_t* buf, int packetBeginIndex, int packetEndIndex, uint8_t receivedCSbyte1, uint8_t receivedCSbyte2)
-    {
-      uint8_t calcChecksum = 0;
-
-      for(uint8_t i=packetBeginIndex + 1; i<packetEndIndex-3; ++i) // packetEndIndex is the "size" of the packet minus 1. Loop from 1 to packetEndIndex-4 because the checksum is calculated between $ and *
-      {
-        calcChecksum = calcChecksum^buf[i];
-      }
-
-      uint8_t nibble1 = (calcChecksum&0xF0) >> 4; //"Extracts" the first four bits and shifts them 4 bits to the right. Bitwise AND followed by a bitshift
-      uint8_t nibble2 = calcChecksum&0x0F;
-
-      uint8_t translatedByte1 = (nibble1<=0x9) ? (nibble1+'0') : (nibble1-10+'A'); //Converting the number "nibble1" into the ASCII representation of that number
-      uint8_t translatedByte2 = (nibble2<=0x9) ? (nibble2+'0') : (nibble2-10+'A'); //Converting the number "nibble2" into the ASCII representation of that number
-
-      if(translatedByte1==receivedCSbyte1 && translatedByte2==receivedCSbyte2) //Check if the checksum calculated from the packet payload matches the checksum in the packet
-      {
-        return true;
-      }
-      else
-      {
-        return false;
-      }
-    }
-
-void parseNMEA(DataFrame* data, uint8_t* buf, int beginIndex, int endMessage) {
-	if (!nmeaChecksumCompare(buf, beginIndex, endMessage, buf[endMessage-2], buf[endMessage-1])) return;
-
-	char nmea[5];
-	int current = beginIndex;
-	for (int i = 0; i < 5; i++) {
-		current++;
-		nmea[i] = buf[current];
-	}
-	char val[20] = {};
-	if (strcmp(nmea, "GPGGA") == 0) {
-		current++;
-		int ind = 0;
-		while (buf[current] != ',') {} //skip Time for now
-		while (buf[current] != ',') {
-			current++;
-			val[ind] = buf[current];
-			ind++;
-		}
-		data->gps.lat = atof(val);
-		ind = 0;
-		memset(val, 0, sizeof val);
-		while (buf[current] != ',') {current++;} //skip N/Z for now
-		while (buf[current] != ',') {
-					current++;
-					val[ind] = buf[current];
-					ind++;
-		}
-		data->gps.lng = atof(val);
-	}
-	if (strcmp(nmea, "GPVTG") == 0) {
-			current++;
-			int ind = 0;
-			for (int i = 0; i < 6; i++) {
-				while (buf[current] != ',') {current++;}
-			}
-			while (buf[current] != ',') {
-						current++;
-						val[ind] = buf[current];
-						ind++;
-			}
-			data->gps.speed = atof(val);
-	}
-}
-
-void parseGPS(DataFrame* data, uint8_t* buf, uint16_t size) {
-	int beginMessage = 0;
-	int endMessage = 0;
-	bool insideMessage = false;
-	for (int i = 0; i < size; i++) {
-		char c = buf[i];
-		if (c == '$') {
-			insideMessage = true;
-			beginMessage = i;
-			endMessage = i;
-		}
-		if (c == '\r' || endMessage - beginMessage >= 80) {
-			insideMessage = false;
-			if (endMessage - beginMessage < 80) parseNMEA(data, buf, beginMessage, endMessage);
-		}
-
-		if (insideMessage == true) {
-			endMessage++;
-		}
-	}
-}
+#include <MTU/MPPT_parsing.h>
 
 bool calculateCheskumMPPT(uint8_t* buf, uint16_t size) {
 	int checksum = 0;
